completing Question Page

+++ Exam_Web.DataLayer/Entities/TestQuestion.cs
@@ -7,7 +7,7 @@
 
 
 
-        public int TestQuestionId { get; set; }
+        public long TestQuestionId { get; set; }
 
 
         public string Title { get; set; }
